Revert "[SELS-TASK][UI] Getting activity log"
This reverts commit 3e342ef8222c4ce22ed75e7b2664ff73ca9b4d53.
