+++ frontend/src/user/profiles/index.css
@@ -4,52 +4,3 @@
   xl:grid-cols-5 lg:grid-cols-3
   md:grid-cols-2 sm:grid-cols-2 sm:gap-10 
 }
-
-/*********** FOR USER DETAILS ************/
-.user-details-content {
-  @apply flex flex-col h-screen pb-20
-  xl:w-4/5 md:w-3/4 xs:w-full
-  md:pt-10 md:px-16 xs:px-2
-  md:overflow-y-scroll
-}
-
-
-/*********** FOR USER INFO ************/
-.side-content {
-  @apply flex flex-col items-center py-5
-  md:h-screen md:border-r-2 md:border-gray-300
-  xl:w-1/5 md:w-1/4 xs:w-full
-}
-
-.user-info-group {
-  @apply flex flex-col w-full items-center mt-3 xs:px-3
-}
-
-.user-name-group {
-  @apply flex justify-between w-3/5 md:flex-col  md:items-center
-}
-
-.buttons-grid {
-  @apply grid grid-cols-2 mt-2 md:w-full xs:w-3/5
-}
-
-.following-button {
-  @apply py-2 flex flex-col md:justify-center md:text-base xs:text-sm items-center
-}
-
-.following-button-text {
-  @apply text-purple-400 md:text-2xl xs:text-lg w-1/3 font-bold
-}
-
-.other-button-group {
-  @apply flex w-full md:flex-col md:my-8 
-  xs:my-3 xs:justify-end
-}
-
-.activity-log-button {
-  @apply items-end md:p-3 xs:p-2 
-  md:text-lg xs:text-sm md:w-full
-  md:bg-transparent xs:bg-primary
-  md:rounded-none xs:rounded-full 
- hover:bg-primary
-}
